created JWT token for users and groups

+++ emi-space-finder/src/services/spaces/DeleteSpaces.ts
@@ -1,6 +1,7 @@
 import { DeleteItemCommand, DynamoDBClient } from "@aws-sdk/client-dynamodb";
 import { DeleteCommand } from "@aws-sdk/lib-dynamodb";
 import { APIGatewayProxyEvent, APIGatewayProxyResult } from "aws-lambda";
+import { hasAdminGroup } from "../shared/Util";
 
 const ID:string = 'id';
 
@@ -10,6 +11,14 @@ export async function deleteSpaces (
   
   //get ID to pass to db query
   if(event.queryStringParameters && (ID in event.queryStringParameters)){
+
+    if(!hasAdminGroup(event)){
+      return {
+        statusCode:401,
+        body:JSON.stringify('Token no authorised for this group')
+      }
+    }
+
     const spaceId = event.queryStringParameters[ID]
 
     const deleteResult = await dbClient.send(new DeleteItemCommand({
